Split terminal state contracts
