+++ Tests/QuillCodeParityTests/ParityGateTests.swift
@@ -1876,6 +1876,25 @@ final class ParityGateTests: XCTestCase {
         XCTAssertFalse(surfaceText.contains("public struct TerminalCommandSurface"), "WorkspaceSurface should not own terminal command rows.")
     }
 
+    func testTerminalStateContractsLiveOutsideEngine() throws {
+        let engineText = try Self.appSourceText(named: "WorkspaceTerminalEngine.swift")
+        let stateText = try Self.appSourceText(named: "WorkspaceTerminalState.swift")
+
+        XCTAssertTrue(stateText.contains("public struct TerminalCommandState"), "Terminal command state should live in the terminal state contract file.")
+        XCTAssertTrue(stateText.contains("public enum TerminalCommandStatus"), "Terminal command lifecycle labels should live in the terminal state contract file.")
+        XCTAssertTrue(stateText.contains("public struct TerminalState"), "Terminal session state should live in the terminal state contract file.")
+        XCTAssertTrue(stateText.contains("struct WorkspaceTerminalExecutionContext"), "Terminal execution context should live beside terminal state contracts.")
+        XCTAssertTrue(stateText.contains("struct WorkspaceTerminalSessionResult"), "Terminal session result should live beside terminal state contracts.")
+        XCTAssertTrue(engineText.contains("enum WorkspaceTerminalEngine"), "Terminal reducer and shell wrapping behavior should remain in the terminal engine.")
+        XCTAssertTrue(engineText.contains("static func localExecutionContext"), "Terminal engine should own local shell wrapping.")
+        XCTAssertTrue(engineText.contains("static func remoteWrappedCommand"), "Terminal engine should own remote shell wrapping.")
+        XCTAssertFalse(engineText.contains("public struct TerminalCommandState"), "Terminal engine should not own command state DTO definitions.")
+        XCTAssertFalse(engineText.contains("public enum TerminalCommandStatus"), "Terminal engine should not own command status DTO definitions.")
+        XCTAssertFalse(engineText.contains("public struct TerminalState"), "Terminal engine should not own terminal session DTO definitions.")
+        XCTAssertFalse(engineText.contains("struct WorkspaceTerminalExecutionContext"), "Terminal engine should not own execution context DTO definitions.")
+        XCTAssertFalse(engineText.contains("struct WorkspaceTerminalSessionResult"), "Terminal engine should not own session result DTO definitions.")
+    }
+
     func testWorkspaceHTMLRendererDelegatesBrowserRendering() throws {
         let htmlText = try Self.appSourceText(named: "WorkspaceHTMLRenderer.swift")
         let browserText = try Self.appSourceText(named: "WorkspaceHTMLBrowserRenderer.swift")
