+++ docs/DECISIONS.md
@@ -135,7 +135,7 @@
 - Runtime issue and execution-context surface contracts live in `QuillCodeRuntimeSurface.swift`, not in the aggregate `WorkspaceSurface.swift`. `WorkspaceSurface` should keep only the composed `runtimeIssue` payload and delegate construction to runtime/execution builders, while renderers consume the shared severity, diagnostic, and execution-context records directly. Future QuillCloud relay or additional remote execution contexts should extend this contract file first instead of adding renderer-local enums.
 - Fork, compact, and cancelled-send title seeding lives in `WorkspaceThreadSeedBuilder`. The workspace model owns thread creation, selection, stores, and top-bar refresh, while the builder owns visible-message filtering, latest-turn seed selection, compact summary formatting, and first-prompt title derivation. Prompt titles split on all whitespace so cancelled whitespace-only prompts cannot create invisible chat titles.
 - SSH Remote terminal continuity uses marker metadata over the existing noninteractive shell stream. The remote wrapper captures baseline env, applies persisted terminal env deltas, runs the user command, then emits final cwd/env markers that are stripped before the entry is stored. This keeps the first implementation compatible with plain SSH while preserving `cd`, `export`, and `unset` across terminal commands.
-- Terminal session state and shell wrapping live in `WorkspaceTerminalEngine.swift`, not inline in `WorkspaceModel`. The workspace model owns async orchestration, project selection, persistence, and top-bar updates; the terminal engine owns state mutation, local/remote command wrapping, marker parsing, cwd persistence, environment deltas, and marker cleanup. This keeps terminal continuity testable as local, SSH Remote, and future relay execution contexts grow.
+- Terminal session contracts live in `WorkspaceTerminalState.swift`, while shell wrapping and lifecycle reducers live in `WorkspaceTerminalEngine.swift`. The workspace model owns async orchestration, project selection, persistence, and top-bar updates; the terminal state file owns command/session DTOs and execution-context payloads; the terminal engine owns state mutation, local/remote command wrapping, marker parsing, cwd persistence, environment deltas, and marker cleanup. This keeps terminal continuity testable as local, SSH Remote, and future relay execution contexts grow.
 - A second Claude CLI design pass emphasized persistent execution context, immediate tool-card feedback, low-drama safety states, and bounded long-output handling. Remote cards and terminal entries should visually inherit their execution context, queued tools should appear immediately, routine safety checks should stay neutral/green or low-amber rather than scary red, and long stdout should keep head/tail plus explicit expansion instead of pushing the user out of the conversation.
 - A follow-up Claude CLI design pass prioritized native-feeling density over decorative chrome: unified local/remote execution-context pills in the composer and cards, thin state accent rails instead of full-card tinting, auto-collapsed successful tool cards, capped long stdout with explicit expansion, command-palette mode prefixes, debounced streaming repaints, and an `NSTextView`-grade composer for large pastes.
 - Active-chat Find is ephemeral UI state. `Cmd+F` opens a transcript-local bar with focused input, counts, previous/next navigation, and active-result highlighting, but the query and selected result are not persisted to thread JSON because they are editor/navigation state rather than conversation state.
